+++ src/extension/sharepoint.js
@@ -17,7 +17,7 @@
  * @param {Window} [win] The window object for testing
  */
 export async function wordHelper(extensionId, sharePointUrl, win = window) {
-  if (win.location.origin === 'https://word-edit.officeapps.live.com' && !win.hlx?.previewListenerAdded) {
+  if (win.location.hostname.endsWith('word-edit.officeapps.live.com') && !win.hlx?.previewListenerAdded) {
     win.hlx = win.hlx || {};
     win.hlx.previewListenerAdded = true;
     chrome.runtime.onMessage.addListener(({ action, url }, { id: senderId }, sendResponse) => {